# docs/openapi.swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "AI Audit System API",
                 "version":  "2.3.0",
                 "description":  "REST API for AI-powered quality audit system with GPT-5",
                 "contact":  {
                                 "name":  "API Support"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3002",
                        "description":  "Development server"
                    }
                ],
...
    "paths":  {
...
                  "/api/fiches/search":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Fiches"
                                                                  ],
                                                         "summary":  "Search fiches by date with status information",
                                                         "parameters":  [
                                                                            {
                                                                                "in":  "query",
                                                                                "name":  "date",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date"
                                                                                           },
                                                                                "description":  "Date in YYYY-MM-DD format"
                                                                            },
                                                                            {
                                                                                "in":  "query",
                                                                                "name":  "includeStatus",
                                                                                "schema":  {
                                                                                               "type":  "boolean",
                                                                                               "default":  true
                                                                                           },
                                                                                "description":  "Include transcription and audit status"
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "List of fiches with status information"
                                                                                   }
                                                                       }
                                                     }
                                         },
                  "/api/fiches/{fiche_id}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Fiches"
                                                                      ],
                                                             "summary":  "Get fiche details with recordings (auto-cached or force refresh)",
                                                             "parameters":  [
                                                                                {
                                                                                    "in":  "path",
                                                                                    "name":  "fiche_id",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                },
                                                                                {
                                                                                    "in":  "query",
                                                                                    "name":  "refresh",
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                    "description":  "Set to \"true\" to force refresh from external API and upsert to database"
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "Fiche details"
                                                                                       }
                                                                           }
                                                         }
                                             },
                  "/api/fiches/{fiche_id}/cache":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Fiches"
                                                                            ],
                                                                   "summary":  "Get cached fiche data",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "in":  "path",
                                                                                          "name":  "fiche_id",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Cached fiche data"
                                                                                             },
                                                                                     "404":  {
                                                                                                 "description":  "Not cached"
                                                                                             }
                                                                                 }
                                                               }
                                                   },
                  "/api/fiches/{fiche_id}/status":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Fiches"
                                                                             ],
                                                                    "summary":  "Get fiche status (transcription and audit info)",
                                                                    "parameters":  [
                                                                                       {
                                                                                           "in":  "path",
                                                                                           "name":  "fiche_id",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Fiche status information"
                                                                                              },
                                                                                      "404":  {
                                                                                                  "description":  "Fiche not found in database"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
                  "/api/fiches/status/batch":  {
                                                   "post":  {
                                                                "tags":  [
                                                                             "Fiches"
                                                                         ],
                                                                "summary":  "Get status for multiple fiches",
                                                                "requestBody":  {
                                                                                    "required":  true,
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "object",
                                                                                                                                            "properties":  {
                                                                                                                                                               "ficheIds":  {
                                                                                                                                                                                "type":  "array",
                                                                                                                                                                                "items":  {
                                                                                                                                                                                              "type":  "string"
                                                                                                                                                                                          }
                                                                                                                                                                            }
                                                                                                                                                           }
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Status information for multiple fiches"
                                                                                          }
                                                                              }
                                                            }
                                               },
...
              },
...
}